bugfix: made ProfileMenu component's
