+++ src/pages/orders.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const OrdersPage = () => {
+  return (
+    <div>
+      <p className="mt-20 mb-20 text text_type_main-medium">Просим прощения. Раздел в разработке.</p>
+    </div>
+  );
+};
+
+export default OrdersPage;
