+++ src/pages/profile.jsx
@@ -3,11 +3,11 @@ import {NavLink, useNavigate} from "react-router-dom";
 import styles from './profile.module.css'
 import {Button, EmailInput, Input, PasswordInput} from "@ya.praktikum/react-developer-burger-ui-components";
 import {useDispatch, useSelector} from "react-redux";
-import {getLogOutUser, getUserInfo, setUserInfo} from "../services/actions/user";
+import {getUserInfo, setUserInfo} from "../services/actions/user";
 import {getCookie} from "../utils/utils";
 
 const ProfilePage = () => {
-  const {userInfo, accessToken, userInfoFailed, userInfoAnswer} = useSelector(store => store.user);
+  const {userInfo, accessToken, userInfoAnswer} = useSelector(store => store.user);
   const [isInfoChanged, setIsInfoChanged] = useState(false);
   const dispatch = useDispatch();
   const navigate = useNavigate();
@@ -20,8 +20,6 @@ const ProfilePage = () => {
   });
 
   const token = getCookie('token');
-  const setActive = ({isActive}) => isActive ? `${styles.link} text text_type_main-medium ${styles.link_active}`
-    : `${styles.link} text text_type_main-medium`;
 
   const editUserInfo = useCallback(e => {
       e.preventDefault();
@@ -40,19 +38,12 @@ const ProfilePage = () => {
     setValue({...form, [e.target.name]: e.target.value});
     setIsInfoChanged(true);
   }
-  const getOut = () => {
-    dispatch(getLogOutUser({token}));
-  }
+
   useEffect(() => {
     if (!token) {
       navigate('/login');
     }
     dispatch(getUserInfo({accessToken}));
-    // if (accessToken) {
-    //   setValue({...form, name: userInfo.name, email: userInfo.email, password: '', accessToken: accessToken})
-    //   // } else {
-    //   //   navigate('/login');
-    // }
 
   }, [accessToken]);
 
@@ -61,25 +52,7 @@ const ProfilePage = () => {
   }, [userInfo]);
 
   return (
-    <main className={styles.wrapper}>
-      <div className={styles.container}>
-        <ul className={styles.menu}>
-          <li>
-            <NavLink to="/profile" className={setActive}
-            >Профиль</NavLink>
-          </li>
-          <li>
-            <NavLink to={"/profile/orders"} className={`${styles.link} text text_type_main-medium`}>История
-              заказов</NavLink>
-          </li>
-          <li>
-            <NavLink className={`${styles.link} text text_type_main-medium`} onClick={getOut}>Выход</NavLink>
-          </li>
-          <li className={"mt-20 "}>
-            <p className={"text text_type_main-default text_color_inactive"}>В этом разделе вы можете
-              изменить свои персональные данные</p>
-          </li>
-        </ul>
+    <>
         {userInfoAnswer
           ? <form className={`${styles.form}`} onSubmit={editUserInfo} onReset={resetUserInfo}>
             <Input type={"text"} extraClass={``} placeholder={'Имя'} value={form.name} name={"name"}
@@ -107,8 +80,7 @@ const ProfilePage = () => {
           ? <p className={`${styles.menu} mt-20 text text_type_main-default text_color_active`}>Вы успешно вышли из
             профиля.</p>
           : null}
-      </div>
-    </main>
+    </>
   );
 };
