+++ src/pages/profile.module.css
@@ -1,44 +1,3 @@
-.wrapper {
-  align-items: center;
-  display: flex;
-  justify-content: center;
-  width: 100%;
-}
-
-.container {
-  align-items: flex-start;
-  column-gap: 60px;
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  margin-top: 120px;
-  max-width: 1280px;
-  width: 100%;
-}
-
-.menu {
-  display: flex;
-  flex-direction: column;
-  list-style: none;
-  margin: 0;
-  max-width: 320px;
-  padding: 0;
-  width: 100%;
-}
-
-.link {
-  align-items: center;
-  color: #8585AD;
-  display: flex;
-  justify-content: left;
-  min-height: 64px;
-  text-decoration: none;
-}
-
-.link_active {
-  color: #F2F2F3;
-}
-
 .form {
   align-items: center;
   display: flex;
